requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
assemblyai==0.33.0
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
cloudpathlib==0.19.0
coloredlogs==15.0.1
confection==0.1.5
Counter==1.0.0
cymem==2.0.8
distro==1.9.0
et-xmlfile==1.1.0
Flask==2.0.1
Flask-SocketIO==5.3.7
flatbuffers==24.3.25
frozenlist==1.4.1
git-filter-repo==2.45.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
humanfriendly==10.0
idna==3.8
imageio==2.35.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.51.0
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyMatting==1.1.12
pyparsing==3.1.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytube @ git+https://github.com/pytube/pytube@a32fff39058a6f7e5e59ecd06a7467b71197ce35
pytz==2024.1
redis==5.0.8
referencing==0.35.1
regex==2024.9.11
rembg==2.0.59
requests==2.26.0
requests-oauthlib==2.0.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spotipy==2.24.0
srsly==2.4.8
sympy==1.13.3
thinc==8.2.5
tifffile==2024.9.20
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
wasabi==1.1.3
weasel==0.4.1
websockets==13.0.1
Werkzeug==2.0.1
wget==3.2
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.8
youtube-transcript-api==0.6.2
